dir.targets: correctly enclose 'Exists' condition in single quotes
According to MSDN (http://msdn.microsoft.com/en-us/library/7szfhaft.aspx), the
single quotes can be left off for "simple alphanumeric strings or boolean values".
This is not the case here (it is a property), so we should use single quotes.

Note that MSBuild ignores this, but it results in a parser error on Mono's xbuild.