Reenable switches that start with a slash on Mono
Until now, switches on Mono only supported the '-' syntax instead of '/'.
This was done for simplicity to be able to determine if the parameter is a path or MSBuild switch.

By using a similar heuristic as introduced in https://github.com/Microsoft/msbuild/pull/67,
we can now use slashes as well. The heuristic checks if the first segment of a string with
slashes is a directory on disk, if yes it's very likely a path. Otherwise it's a switch.

Fixes #48