[tests] Fixes ResolveAssemblyReference tests
Changes instances of hardcoded paths such as:

`c:\WINNT\Microsoft.NET\Framework\v2.0.MyVersion`

to use `readonly` fields which are constructed to be valid on both
Windows and Unix. Replace all occurrences of these paths, to use the
new fields.

Also, adds a s_rootPathPrefix field that can be "C:\\" or "/".

PR-URL: #111