[xaml] TaskParser.Parse: Handle detection of content or file on Unix.
TaskParser.Parse(string contentOrFile, string)
.. first argument can either be a filename or xaml string. Existing
code depended on Path.GetFullPath returning null for xml string, for
this detection. But on Unix, xml string could be a valid file path.

So, instead we utilize the knowledge that:

On Windows:
- xml string will be an invalid file path, so, Path.GetFullPath will
  return null
- xml string cannot be a rooted path ("C:\\<abc />")

On Unix:
- xml string is a valid path, this is not useful as Path.GetFullPath
  will return !null in most cases
- xml string cannot be a rooted path ("/foo/<abc />")

We have an additional check for rooted paths, just to be able to throw
the correct error in case of:
- Unix, @contentOrFile being a non-existant file path

This fixes all the XamlTaskFactory tests on Unix.