CommandLineBuilder.AppendSwitch: Remove use of FixCommandLineSwitch .
CommandLineBuilder.FixCommandLineSwitch changes a switch of the form
`/foo` to `-foo` on !Windows. Deciding which form is required should
be done by the tool task (eg. Csc) itself, and not in common code for
all tasks using this API.

This patch also removes the use of these in the related tests, such that
no additional tests fail.