Use the xunit Visual Studio runner
This allows test discovery, execution, and (most importantly) debugging
from within Visual Studio 2015.

Because of OSS signing, the assemblies won't load in xunit without setting
shadowCopy=false.  To accomplish that for the VS runner I created a
(global) xunit.runner.json settings file and ensure that it's copied to
the output directory before the build returns.

Since there was already some needed-before-tests-run stuff, I consolidated
it with the xunit config placement in src/CopyTestPrerequisites.proj.
That's referred to from every test project, so we don't have race
conditions or dependencies on build order.