Use the default location for NuGet package restore
This allows NuGet to follow its instincts and restore packages to a
non-repo-specific location, for the most part.

To work around problems with needed-at-test-runtime assemblies not being
copied locally as part of the test-project build (because NuGet won't
generate the right .props file for our project.json), I'm also
(redundantly) restoring xunit packages to the repo-specific path, so I can
copy the assemblies from a known location to bin.