Tell VS test runner to reduce parallelism
The default settings for xunit are to parallelize things at a much
higher rate than MSBuild's tests can handle.  Set the runner
configuration to a) only do one assembly at a time (this is the default
but I'm setting it explicitly as documentation that we require it), and
b) only use one thread when within an assembly to run tests.

While I was reading the documentation for the configuration, I also
found the "methodDisplay" setting, which allows us to go back to the
MSTest-style method-name-only display in the test explorer window.