Restore NuGet packages from build.proj
When we moved package restore from build.cmd to build.proj, we stopped
restoring the packages required for "normal" reasons, as opposed to build
tools.

This adds a step that will restore the packages from the entry point
project.

It's unfortunate that nuget doesn't understand traversal projects, which
means restoring packages based on the .sln rather than the list in
build.proj which will actually be built.  We could potentially also
restore **\project.json, but then that would require N calls to nuget.exe.
This seems better for now.