Disable FindFileOnPathAfterResizingBuffer test
This test manipulates our internal representation of MAX_PATH, which
caused us to call shell32!SHGetFolderPathW with a tiny buffer, which then
overran its buffer, corrupting memory and crashing corerun.

We need a better way to test long-path behavior.  Changing MAX_PATH is not
a good idea.  Adding a new test for this behavior is #236.