Introduce ToolsetDefinitionLocation.Default
Previously, the "default" ToolsetDefinitionLocation settings were scattered across the tree as Registry | ConfigurationFile (or vice-versa).  Unifying those into a build-in enum member gives us a nice way to centralize changes . . .

Which we want to do for netcore, because netcore does not support reading from the registry (doesn't exist cross-platform) or the configuration file (netcore doesn't support those APIs).  This was aready done in xmake.cs, but that change didn't affect unit tests, many of which were failing because they couldn't find the current toolset version (14.1) in the registry--where they didn't need to be looking!