Scope AppDomain flagging for GetTaskFactory
The previous use of FEATURE_APPDOMAIN caused netcore (non-appdomain) builds to always attempt to use an AssemblyTaskFactory.  That was confusing for some unit tests I was looking at that were trying to use CodeTaskFactory, and failed with a can't-load-that-from-an-assembly error.

Now, we check the same conditions but throw an explicit exception in the unsupported-because-we-don't-have-appdomains case.