Flag tests that use CodeTaskFactory incidentally
These tests use CodeTaskFactory to get a task that behaves in a certain
way but don't actually test its functionality.  Flagging them behind
FEATURE_CODETASKFACTORY.  They could probably be rewritten to work without
it (perhaps by having a test-tasks assembly that is built as part of the
normal build).  Filed #248 to track that.