Adjust paths returned by property functions only as strings
Works around #249 for unit tests, but a deeper fix for collections is
probably required.

The ToString() call forced any value returned from a property function
into a string, which was the root cause of the test failures--the string
returned was a representation of a generic that didn't stringize with high
precision.