Flag away code that depends on installation
In NetCore, there's not really a good concept for the "DotNetFrameworkPath"--the framework comes in piecemeal from packages, and we'll rely on NuGet to get the right bits for a project that's building using NetCore MSBuild.  Disabling code that checked for installation and used `typeof(object).GetTypeInfo().Module` to find the framework directory.