Do not throw on assembly-load failure in Expander
When we started moving to CoreCLR, we had to replace an
Assembly.LoadWithPartialName call with Assembly.Load, but the exception
behavior differs between those two methods.  Load throws when no assembly
matching the given name is found, which caused attempts to load property
functions to have an unexpected failure mode.

To keep behavior similar on both codepaths, I'm swallowing
FileNotFoundException from Assembly.Load--the candidate assembly will just
continue to be set to null, and the old error-handling code will apply.