Match CI build version to Debian Package Build Version.
The Debian Package should have the same version as the rest of the packages
being created as a part of the full build. These changes enable that.

NOTE: The debian package revision is unused but still included in the package
to allow for hyphens in the package version. (https://www.debian.org/doc/debian-policy/ch-controlfields.html#s-f-Version)