Improve diagnosability of LogBuildStarted test
This test seems a bit flaky; it's failed a few times in CI builds for no
discernable reason (and passed on retry of the same source).
Unfortunately, the failure was just that an Assert.True failed, which
didn't provide much actionable information (not even which of the two
asserts it was).

In the hopes of making the next failure clearer, this change:

 * Splits the two asserts into two tests so we can see which is failing
 * Moves to Assert.Equals, to get xunit's stringification of the objects
   on failure (so we can hopefully see a difference).