Use system dependencies on system builds
When we build msbuild using the system installed msbuild, all the
involved build dependencies (DLLs, EXEs, target files) should be system
dependencies, not a mix between system dependencies and app local ones.

This fix deletes the import in the dir.targets which overrides the
dependency location regardless of the build. The consequence of this
import was that the system build used the program files
Microsoft.CSharp.Core.Targets but the app local CSC task. On VS update 1
this breaks the build