Invert conditions for passing /nr to child
Previously, every child node was started with implicit
`/nodeReuse:true`.  It was either passed explicitly (by adding `/nr` in
the code that's changing with this commit), or implicitly (because
`/nr:true` is the default).

That wasn't a problem because the main node knew whether or not to shut
down it's children (*it* had the correct state).  But since I'm adding
reuse eligibility to the host-client handshake, the child now needs to
know the state correctly.

While I was in there, I used the full, spelled-out version of the option
because short options are for humans, not programs.