Changing native out flag to output flag to match compiler. Also changing the rsp file spliting to handle both /r/n and /r
Fixing the input path to the native compiler. It was adding /native to it before.

Switching compile native to System.CommandLine for args parsing and removing quotes from -native.rsp.

Fixing build_appdeps.cmd: Removed the * from __ApPDepSDK as xcopy does not honor that. In order to stay independent of the version of appDep, we CD into the appDepSDK and then navigate to the first folder we find, which should always be one, as we clear the folder before restoring appDep, after that we xcopy