Restore using project.json files instead of solution file
We are still using full framework NuGet.exe, and if we try to do a restore with
it based on the solution file while building with .NET Core MSBuild, our
projects will be evaluated as full framework projects, with imports from the
full framework build tools directory, which may not have been restored or up to
date.  So just call NuGet restore using the project.json files.