[test] Fix tests running on mono by running them ..
.. in app domains.

Rationale:

The test assemblies, along with the newly built msbuild assemblies are in
*_Deployment_Test dir. Xunit is run from it's nuget packages dir.
When Xunit is run with the test assembly, mono has to fall back to GAC
to resolve the msbuild assemblies. But running tests against mono's
msbuild assemblies would be incorrect.

This manifests in the output as (one instance):

      System.MissingMethodException : Method 'TaskFinishedEventArgs..ctor' not found.
      Stack Trace:
          at (wrapper managed-to-native) System.Reflection.MonoMethod:InternalInvoke (System.Reflection.MonoMethod,object,object[],System.Exception&)
          at System.Reflection.MonoMethod.Invoke (System.Object obj, BindingFlags invokeAttr, System.Reflection.Binder binder, System.Object[] parameters, System.Globalization.CultureInfo culture) <0x10ea3db80 + 0x000a3> in <filename unknown>:0

If we make Xunit run the tests in app domains, then when loading the
test assembly for the app domain, mono is able to find the (correct)
newly built msbuild assemblies and the tests can run correctly.

So, don't use -noappdomain option for mono builds.