Update NuGet references to rc2-23712
This should enable us to run on Mac OS X by using a newly-available API.
This package is the latest available on myget minus a couple of revisions
to account for an already-fixed but not published CoreFX bug. A couple of
packages I manually bumped to the latest available to resolve build
errors.