Revert "Copy .mdb files when they are generated by the compiler instead"
This reverts commit 658fc55ea8af591299d7d59bc7a4fd3908b39f6d.

That change detected "is building using the Mono compiler" through the
condition "is not on Windows_NT", which isn't accurate since the Roslyn
compiler works fine on Linux and OS X now. That caused errors to find
`.mdb` files when building with Roslyn and creating debug symbols on
Unices--so this fixes #459.