Add a workaround for a Sockets issue on Unix.
Adding an explicit reference to System.Net.NameResolution in our dotnet-projectmodel-server.Tests.  The current build of Roslyn references an old version of System.Net.NameResolution which is the version that gets used during the test.  Explicitly referencing the current CoreFX version for now.  This should be reverted when the underlying bug gets fixed, or Roslyn moves to a newer CoreFX version.