Add support for resolving $(MSBuildExtensionsPath) with fallback search paths
In Mono's xbuild, $(MSBuildExtensionsPath) is $mono_prefix/lib/mono/xbuild.
On OS X, this is:

    `/Library/Frameworks/Mono.framework/Versions/Current/lib/mono/xbuild`

Consequently, when resolving an Import for
$(MSBuildExtensionsPath)\Foo\Abc.targets, the following path will be checked:

    `/Library/Frameworks/Mono.framework/Versions/Current/lib/mono/xbuild/Foo/Abc.targets`

This works, but has a problem when upgrading the Mono.framework package:
a *new* /Library/Frameworks/Mono.framework/Versions/@VERSION@ directory
is created, so anything installed into the previous directory will be
"lost". The file is still there, but xbuild will no longer find it.

To address this upgrade scenario, Mono's xbuild also checks a fallback path:

    `/Library/Frameworks/Mono.framework/External/xbuild`

This location is not within a versioned framework directory, and thus files
installed into it will survive Mono.framework package upgrades.

Unfortunately, this means that $(MSBuildExtensionsPath) on xbuild is no longer
a single directory; it is a set of directories, and xbuild will check each
directory to see if the file exists:

/Library/Frameworks/Mono.framework/Versions/Current/lib/mono/xbuild/Foo/Abc.targets
/Library/Frameworks/Mono.framework/External/xbuild/Foo/Abc.targets
...

Mono would like to eventually migrate to MSBuild instead of xbuild, which
requires bringing xbuild's fallback support to MSBuild.

Implementation details follow:
-----------------------------

With this feature, MSBuildExtensionsPath* when used in an Import like

    <Import Project='$(MSBuildExtensionsPath)\Foo\Abc.targets' />

will cause a lookup for the relative file path `Foo\Abc.targets` in the location pointed
to by the $(MSBuildExtensionsPath) property. But if it cannot be found there,
then it will use some fallback paths to look for that file. So, for such an
import the targets file is searched using the following paths (in order):

  1. The location from the MSBuildExtensionsPath* property

  2. Search paths found in the Toolset, which currently gets populated from
     app.config, like:

        <msbuildExtensionsPathSearchPaths>
            <searchPaths os="osx">
                <property name="MSBuildExtensionsPath" value="/Library/Frameworks/Mono.framework/External/xbuild/;/tmp/foo"/>
                <property name="MSBuildExtensionsPath32" value="/Library/Frameworks/Mono.framework/External/xbuild/"/>
            </searchPaths>
        </msbuildExtensionsPathSearchPaths>

     On OSX, this will use the two paths - `/Library/Frameworks/Mono.framework/External/xbuild/` and
     `/tmp/foo` as the fallback paths to look for `Foo\Abc.targets` .

This is available only when used in an Import element and it's Condition. So,
the following common pattern would work:

     <Import Project="$(MSBuildExtensionsPath)\Foo\Abc.targets" Condition="'Exists('$(MSBuildExtensionsPath)\Foo\Abc.targets')'" />

The value of the MSBuildExtensionsPath* property, will always be "visible" with
it's actual value, like, when read or referenced anywhere else. This is a very
limited support, so, it doesn't come in to effect if the explcit reference to
the $(MSBuildExtensionsPath) property is not present in the Project attribute
of the Import element. So, the following is not supported:

     <PropertyGroup>
        <ProjectPathForImport>$(MSBuildExtensionsPath)\Foo\Abc.targets</ProjectPathForImport>
     </PropertyGroup>
     <Import Project='$(ProjectPathForImport)' />

So, effectively, any targets file which would earlier have been imported
from the location specified by MSBuildExtensionsPath* property, would continue
to be imported as is. But for any imports requested which are not found there,
extra search paths will be used to find those.