Fill out pkg installer targets
This causes three pkg files to be produced:

* A component pkg for the shared framework.
* A component pkg for the shared host.
* A product archive which includes the above two components.

The product archive also needs a distribution.xml file which contains
metadata about the package (name, title, images, etc.).

The installer for the "SDK" itself is still using logic implemented in
package-osx.sh. We should move this logic into the build target as well,
but we may want to wait until the CLI is using the shared framework to do so.