Add pre-reqs doc to the repo (#2219)
* Add pre-reqs doc to the repo 

This document outlines all of the native dependencies that are needed to run the CLI tools for Ubuntu, Windows, CentOS and OS X. 
As per chat with @leecow documentation does not need special approval to get merged. 

skip ci please

* Update cli-prerequisites.md

* Responding to PR feedback

I've changed the Ubuntu reqs to be the same as CentOS pre-reqs just with the correct package names. This seems like a more sane thing to do.

* Update cli-prerequisites.md