Exit when build tools can't be restored instead of continuing
When build tools restore failed, we would just continue before which would lead to all sorts of errors.
Instead we now cleany exit with an error code. We also need to remove the $__PROJECT_JSON_FILE or it would skip
the build tools restore the next time we retry after getting an error.