Skip CI test result archiving if no test result files found
We can't differentiate PR jobs by the targeting branch.
This is an issue because PRs targeting master should only have Windows
CI jobs.

For tests, this means the CI should only collect test results when test
result files are there. On failing tests, the main build will fail which
will also fail the job.

The only risk here is that if someone turns off the tests entirely, the
main build succeeds and Jenkins will not complain.