Convert NotExpectedException to exception filters
There was a pretty common pattern across the codebase:
* try an I/O operation
* If it succeeds, great
* If it threw an IO exception, log a specific error or handle it in some
  other way
* If it threw some *other* exception, rethrow it.

Conveniently, this is now supported by the "exception filters" language
feature, new in C# 6. This approach is less code, clearer in intent, and
has some performance and debuggability benefits.

I did the conversion by find-all-references on
`ExceptionHandling.NotExpectedException` and `IsIoRelatedException`. In
almost all cases, the latter positively-named form was a better fit for
the use.