Read symlink target with GetLastWriteFileUtcTime
Fixes #322. If the path passed to
`NativeMethods.GetLastWriteFileUtcTime` is a symlink, MSBuild will now
by default return the last write time of the symlink's *target*, not the
symlink itself.

The old behavior returned the last write time of the symlink itself,
which is wrong because it violates the rule of incremental builds: we
should only skip a target if we're confident that running the target
would produce the same result we already have. Since running the target
will probably read the file (through the symlink), the new behavior is
more correct.

The escape-hatch environment variable `MSBUILDUSESYMLINKTIMESTAMP` is
introduced in case this causes unforeseen regressions. Normally I
wouldn't introduce a go-back-to-buggy-behavior flag, but since timestamp
up-to-date checks are a deep part of MSBuild infrastructure, we're a bit
nervous about changing their behavior entirely.