Managed changes for FileTracker->FileTrackerXX
Port of internal VS repo commit 3071fdfce.

MSBuild provides a public API for divining the appropriate paths to
Tracker.exe and FileTracker.dll. This change returns a copy that is
appropriate now after the Detours-required rename to FileTracker32.dll and
FileTracker64.dll.

There is much confusion here.  First, there are now two DLLs where once
there was one. Fortunately, the native Detours code transparently handles
slecting the right DLL given a path to *either*. So I'm just returning the
32-bit version. That means that we no longer need to have the caller
indicate what type of tool will be called so that we can dig around for
the right binary. But that was part of our public interface, so I left the
parameter and now just ignore it.

Second, there was a lot of logic to find the right path to tracker.exe.
From spelunking through history, I *think* this is because tracker.exe
isn't shipped with the Framework version of MSBuild.  It is shipped in the
.NET SDK and with Build Tools.  Now that Build Tools exists, I think
that's the only place that it's needed, and the old tracker.exe isn't
compatible with the new filetracker.dll, so I removed the fallback logic:
it is now found within the MSBuild toolset, or not at all.