Simplify ForceOutOfProcTracking for cross-bitness
Port of internal VS repo change 4543e6aa6.

This is a documented public API so I left it around with the same
signature even though I couldn't find any consumers. The logic to detect
when a tool will run with a different architecture so that a caller can
launch the appropriate tracker.exe for the tool is now obsolete; if
FileTracker is loaded in-proc we can trace a child process no matter its
bitness.