Remove File.Copy optimization during expansion
Previously we'd keep track of any file that we extracted once and try
to reuse that file (by copying it) if we needed the same file later at a
different destination.  The reason was that it's theoretically faster to
a file copy than a createfile and write, since the copy can happen
entirely in the kernel.  In practice we were foiled by AV scanners.

This happens to be the only time during extraction where we let a file
close after writing it and then try and use it again.  Sure enough on
fast machines we were seeing that as soon as we closed it MsMpEng would
map the file for exclusive access causing our copy to fail with a
sharing violation.

To fix this, I've removed the copy optimization and will just copy the
file from the in-memory archive every time.