Archiver should not keep files Open
Previously I attempted to save the cost of reading/decompressing files
by keeping a copy in memory.  That resulted in too much memory pressure
so I later swapped in temporary files.  This defeated the initial goal
of keeping the streams  alive to some extent since the files would
be flushed to disk.

In practice since all the nupkgs we are packing also have raw files so
we weren't even saving much CPU by avoiding a second decompression.

Switch the archiver to just first index all files and save the source
information to read at a later point.  After this when building the
archive we'll reopen the files/zips and copy from there.

My measurements show that this actually improve the archiving perf and
the lack of temp files means we don't hit the ulimit restriction on OSX.