Moving certain documents to the dotnet/core repo (#3660)
* Moving certain documents to the dotnet/core repo

Moving the following documents to the dotnet/core repo, to make maintenance easier and also to keep all of the various support documents for .NET Core all-up in the same place. 

The documents changed:

* supported-os-matrix.md
* known-issues.md
* cli-prerequisites.md


Also editing the README.md document to point to the new repo and its documents where needed.