Update path to versionless registry for GenericBootrsapper
From Dev15 onwards this value should be common for all VS versions and it will
be outside of the VS hive.

Until now Software\Wow6432Node\Microsoft\GenericBootstrapper\$(VSVersion)
was used with value name Path and it was pointing to "C:\Program Files
(x86)\Microsoft Visual Studio $(VSVersion)\SDK\Bootstrapper".

From now Software\Wow6432Node\Microsoft\GenericBootstrapper will be used
with value name Path and pointing outisde of VS hive to "Program
Files(x86)\Microsoft SDKs\ClickOnce Bootstrapper".

This is required because this will be set as a singleton MSI and it needs
to be outside of the VS hive.