Handle missing terminal capabilities.
When probing for terminal capabilities, tput will return exit-code 1
when the attribute is not valid for the current terminal.

If we don't handle that exit-code, the shell-script too will exit with
failure.

This patch simply ignores errors for attributes we cannot use,
resulting in the same result as if we never probed for it in the first
place.