Add support for assembly resolution from config file The code is mostly copied from AssemblyFoldersEx and follows most of the same logic. I didn't implement all the features and options from that class as they don't seem to be used/needed.
In Dev15, Visual Studio setup will drop this file into MSBuild folder in
order for Add References dialog and Build to find reference assemblies.
This will allow for projects to reference, say Microsoft.Build.dll, without
providing a hint path. Previously this was done via the registry but Dev15
will no longer write those entries in all cases. When they are not in the
registry, MSBuild will use this file.