Look for dependencies next to task assemblies
Fixes #658.

Assume we have an assembly called Task.dll that contains some number of tasks
a project may wish to use, and that Task.dll depends on Dependency.dll.
Currently .Net Core version of MSBuild can load Task.dll and run the tasks
within it, but will fail to find and load Dependency.dll unless it is located
immediately next to msbuild.exe.

The fix here is to update `CoreCLRAssemblyLoader` (our implementation of the
`AssemblyLoadContext` type) to remember the directories of loaded task
assemblies and to search those directories when trying to satisfy a dependency.