Copy shim targets only on targeting full framework
These shim targets are designed as a compat shim for SDKs that *used* to
be part of MSBuild proper but have now moved. They're thin redirects to
the new canonical targets-file location. d743a48c renamed them to their
final names, so they no longer have `.Shim` in their filenames. That
simplifies many things, but triggers the existence checks in
Common.targets and cause them to be included. Since they redirect to paths
that don't make sense cross-platform, they shouldn't be packaged next to
MSBuild-for-.NET-Core.

In the long term, these shims *should* be deployed with full-framework
MSBuild, so they'll probably need to make a reappearance in a tfm-specific
way in the nuspec.