Avoid IsNullOrWhitespace for CLR 2
`String.IsNullOrWhitespace` was added in .NET 4.0 so it's not available in
MSBuildTaskHost, which targets 3.5. Avoiding it in shared code.

A string that's all whitespace will fail the subsequent checks, so the
check in MaybeAdjustFilePath is only a slight optimization in Unix
environments.

The use in GetParentProcessId is less trivial but not needed for
MSBuildTaskHost, since there's no xplat .NET 3.5.