Set codeBase for MSBuild assemblies in x64 MSBuild (#836)
Initial fix for #834. Setting `codeBase` to ensure that all MSBuild
assemblies get loaded from the same location (in the MSBuild bin folder)
even if they're loaded from the amd64 MSBuild.exe.

This isn't a great solution--the app.config files should be generated from
a common source, because they should be identical except for the codeBase.
Also, we probably shouldn't even ship copies of the assemblies into the
amd64 folder if they'll never be used. But this is an urgent bug so I'm
providing a quick workaround to be followed up with a long-term fix.