Make vsix use the new amd64 app.config
19bae02 created distinct MSBuild.exe.config files for x86 and x64
MSBuilds, but the .swr file manifest continued to place the x86 version
twice (in bin/ and bin/amd64/). Update it to use the x64 version for the
amd64 directory.

Currently MSBuildTaskHost.exe.config is identical in both versions but I'm
preemptively changing it now in case that becomes untrue in the future.