Remove invalid check in CommandLineBuilder (#815)
When escaping quotes, a check was done to ensure that the number of quotes that will be escaped is even. This doesn't make sense, since there's plenty of command line parsers (including Windows's own CreateProcess) that correctly process any number of embedded quotes (as long as they're escaped properly).

The effects of this check can be seen, by the way, by attempting to define a macro in the project properties for a C++ file which has in its body an odd number of quotes (e.g. a string with an embedded quote). Add `FOO="\""` to the C++ Preprocessor Definitions in the property pages and observe the MSB6001 error that results when you attempt to build.

Replaced obsoleted unit test with one that verifies an odd number of quotes is OK, and removed orphaned string resource.