Add a couple of tests around loading assembly dependencies
Related to #658.

This change introduced a couple of basic tests to verify that we can find a
task assembly's dependencies when they are located in the same directory. In
order to do this, it adds a couple of projects to produce assemblies just for
this test: TaskWithDependency.dll, and Dependency.dll.