Update to xunit.netcore.extensions 629-09
Microsoft.Build.Engine.UnitTests.dll was failing to load on Desktop with
the error:

System.IO.FileNotFoundException: Could not load file or assembly
'System.Runtime, Version=4.1.0.0, Culture=neutral,
PublicKeyToken=b03f5f7f11d50a3a' or one of its dependencies. The system
cannot find the file specified.

This is usually because an assembly targets the wrong .NET version. The
only reference to System.Runtime was transitively through
xunit.netcore.extensions.dll, so I'm updating that reference to one that
works. It's not the current version, because that now depends on a
prerelease xunit.core and I don't want to update that right now too.