Update DotNet CLI and BuildTools (#835)
* DotNet CLI from 1.0.0-rc2-002547 to 1.0.0-preview2-003121 (RTM)
* BuildTools from 1.0.25-prerelease-00410-01 to 1.0.26-prerelease-00627-02 which contains our changes with NuGet pack properties
* Update init-tools.cmd to use the new URL for downloading DotNet CLI
* Update all references to corerun to dotnet instead.
* Update build.cmd to pass the OverrideToolPath property so that tasks will run under dotnet.exe
* Update to Newtonsoft.Json 9.0.1
* Update compilers to fix issues with older compiler and newer dotnet CLI

I had to remove the usage of our "bootstrapped" dotnet cli because the runtimeconfig.json file is telling .NET to run using a "shared" framework. During the 2nd build, we use a bootstrapped dotnet.exe but it is unable to find the shared framework. This will need to be addressed going forward and we'll need to decide if we are going to be a stand-alone app or use a shared framework. The implication here is that other processes like csc.exe will need to use the same framework that MSBuild uses.