Do not require msbuild.exe.config on .NET Core
All of the code that could read configuration data out of
msbuild.exe.config is dependent on System.Configuration, which isn't
available in .NET Core. The reading code was all #if'ed out, but we still
required that the file be physically present, or MSBuild would fail on
initialization.

I made `IsValidMSBuildPath` dependent on the same compile-time check as
src/Shared/ToolsetElement.cs.