Open source the XMakeConversion project
Migrate the XMakeConversion project from the Visual Studio sources to the
MSBuild repo. This is a legacy project and we do not expect to do any further
development on it, nor accept any pull requests pertaining to it. However, we
still need it for the moment so it may as well live alongside the rest of the
MSBuild sources.

This commit just copies the project and sources files as-is. Future commits will
properly integrate it into the build and packages.