Open-source the OrcasEngine project
Migrate the OrcasEngine project from the VisualStudio sources to the MSBuild
repo. This is a legacy project and we do not expect to do any further
development on it, nor accept any pull requests pertaining to it. However, we
still need it for the moment so it may as well live alongside the rest of the
MSBuild sources.

This commit just copies the project and source files as-is. Future commites will
property integrate it into the build and packages.