"Publish" our tasks assembly into platform-specific folders inside our NuGet package.
This allows us to put platform-specific dependencies in the NuGet package, and fixes the issue of getting FileNotFound 'System.Security.Cryptography.Algorithms' on net46 by adding that assembly to our NuGet package.

Along with this change, I refactored the nuspec to include all files under the "PackagesLayout" folder, which means we no longer have to duplicate each file.  "Build" just needs to place the file in the correct spot under "PackagesLayout".

Fix #45