Avoid node reuse during desktop-hosted builds (#896)
If msbuild is invoked with the (default) `/nodeReuse:true` argument,
worker nodes hang around for a period of time after the build is complete
to see if more work will be incoming. This is normally a nice
startup-time-reducing perf optimization. But in our own build, it means
that many newly-created MSBuild assemblies are opened and uneditable,
failing builds and causing cleans to fail.

Changing the Windows build script to just always pass `/nodeReuse:false`
when not using the CoreCLR MSBuild.