Update build to produce NuGet packages
Update the build to produce NuGet packages.

We re-use the packaging capabilities from Microsoft.DotNet.BuildTools.
However, we want to use the version produced by NerdBank.GitVersioning,
rather than than that produced by versioning support on
Microsoft.DotNet.BuildTools. This requires defining and wiring up a
couple of new targets so that the necessary properties and items are
defined when the packaging occurs.

Also, it was necessary to define the PackagingTaskDir property. Most of
the targets in Microsoft.DotNet.BuildTools use the BuildToolsTaskDir
property to find the assemblies containing their tasks, but packaging is
a little different. Without this, we would end up trying to load the
netcore version of the packaging tasks, rather than the desktop version.